Api key managment

+++ app/controllers/api/base_controller.rb
@@ -1,24 +1,40 @@
 module Api
   class BaseController < ActionController::Base
+    include Pundit
+
     skip_before_action :verify_authenticity_token
 
     before_action :authenticate
+
+    after_action :verify_authorized, except: :index
+
+    after_action :verify_policy_scoped, only: :index
 
     attr_reader :current_user
 
     helper_method :parent, :collection, :resource, :current_user
 
+    def show
+      authorize(resource)
+    end
+
     def create
       build_resource
+
+      authorize(resource)
 
       resource.save!
     end
 
     def update
+      authorize(resource)
+
       resource.update!(resource_params)
     end
 
     def destroy
+      authorize(resource)
+
       resource.destroy!
 
       head :ok
@@ -40,12 +56,32 @@ module Api
       render :exception, status: :not_found
     end
 
+    rescue_from Pundit::NotAuthorizedError do
+      head :forbidden
+    end
+
     private
 
     def authenticate
-      authenticate_or_request_with_http_token do |token, options|
+      authenticate_or_request_with_http_token do |token, options| # rubocop:disable Lint/UnusedBlockArgument
         @current_user = User.find_by(token: token)
       end
+    end
+
+    def resource
+      raise NotImplementedError
+    end
+
+    def resource_params
+      raise NotImplementedError
+    end
+
+    def build_resource
+      raise NotImplementedError
+    end
+
+    def collection
+      raise NotImplementedError
     end
   end
 end
